Refactor old OrphanageData to leave fields until image
