+++ packages/mobile/src/pages/CreateOrphanage/DataSectionOne/styles.ts
@@ -1,0 +1,25 @@
+import styled from 'styled-components/native'
+
+export const ImagesInput = styled.TouchableOpacity`
+  background-color: rgba(255, 255, 255, 0.5);
+  border-style: dashed;
+  border-color: #96d2f0;
+  border-width: 1.4px;
+  border-radius: 20px;
+  height: 56px;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 32px;
+`
+
+export const UploadedImagesContainer = styled.View`
+  flex-direction: row;
+`
+
+export const UploadedImage = styled.Image`
+  width: 64px;
+  height: 64px;
+  border-radius: 20px;
+  margin-bottom: 32px;
+  margin-right: 8px;
+`
