--- packages/mobile/src/pages/CreateOrphanage/OrphanageData/index.tsx
@@ -1,148 +1,0 @@
-import React, { useCallback, useState } from 'react'
-
-import { Feather } from '@expo/vector-icons'
-import { RouteProp } from '@react-navigation/core'
-import { StackNavigationProp } from '@react-navigation/stack'
-import { launchImageLibraryAsync, MediaTypeOptions } from 'expo-image-picker'
-import { ImageInfo } from 'expo-image-picker/build/ImagePicker.types'
-import { Switch } from 'react-native'
-import Toast from 'react-native-simple-toast'
-
-import { CreateOrphanageParamList } from '~/routes/types'
-import api from '~/services/api'
-
-import {
-  Container,
-  Title,
-  Input,
-  ImagesInput,
-  Label,
-  NextButton,
-  UploadedImagesContainer,
-  UploadedImage,
-  NextButtonText,
-  SwitchContainer
-} from './styles'
-
-interface OrphanageDataProps {
-  navigation: StackNavigationProp<CreateOrphanageParamList, 'DataSectionOne'>
-  route: RouteProp<CreateOrphanageParamList, 'DataSectionOne'>
-}
-
-export const OrphanageData: React.FC<OrphanageDataProps> = ({
-  route,
-  navigation
-}) => {
-  const [name, setName] = useState('')
-  const [about, setAbout] = useState('')
-  const [instructions, setInstructions] = useState('')
-  const [openingHours, setOpeningHours] = useState('')
-  const [openOnWeekends, setOpenOnWeekends] = useState(false)
-  const [images, setImages] = useState<string[]>([])
-
-  const { position } = route.params
-
-  const handleSelectImages = useCallback(async () => {
-    const result = await launchImageLibraryAsync({
-      allowsEditing: true,
-      quality: 1,
-      aspect: [5, 4],
-      mediaTypes: MediaTypeOptions.Images
-    })
-
-    if (!result.cancelled) {
-      const { uri: image } = result as ImageInfo
-      setImages(prevState => [...prevState, image])
-    }
-  }, [])
-
-  const handleCreateOrphanage = async () => {
-    const { latitude, longitude } = position
-    const data = new FormData()
-
-    data.append('name', name)
-    data.append('about', about)
-    data.append('instructions', instructions)
-    data.append('opening_hours', openingHours)
-    data.append('latitude', String(latitude))
-    data.append('longitude', String(longitude))
-    data.append('open_on_weekends', String(openOnWeekends))
-
-    images.forEach((image, index) => {
-      data.append('images', {
-        name: `image_${index}.jpg`,
-        type: 'image/jpeg',
-        uri: image
-      } as any)
-    })
-
-    api
-      .post('/orphanages', data)
-      .then(() => {
-        Toast.show('Orfanato cadastrado com sucesso.')
-        navigation.navigate('DataSectionTwo')
-      })
-      .catch(err => {
-        console.log(err)
-      })
-  }
-
-  return (
-    <Container contentContainerStyle={{ padding: 24 }}>
-      <Title>Dados</Title>
-
-      <Label>Nome</Label>
-      <Input value={name} onChangeText={setName} />
-
-      <Label>Sobre</Label>
-      <Input
-        multiline
-        style={{ height: 110 }}
-        value={about}
-        onChangeText={setAbout}
-      />
-
-      {/* <Label>WhatsApp</Label>
-      <Input /> */}
-
-      <Label>Fotos</Label>
-
-      <UploadedImagesContainer>
-        {images.map(image => (
-          <UploadedImage key={image} source={{ uri: image }} />
-        ))}
-      </UploadedImagesContainer>
-
-      <ImagesInput onPress={handleSelectImages}>
-        <Feather name="plus" size={24} color="#15B6D6" />
-      </ImagesInput>
-
-      <Title>Visitação</Title>
-
-      <Label>Instruções</Label>
-      <Input
-        multiline
-        style={{ height: 110 }}
-        value={instructions}
-        onChangeText={setInstructions}
-      />
-
-      <Label>Horário de visitas</Label>
-      <Input value={openingHours} onChangeText={setOpeningHours} />
-
-      <SwitchContainer>
-        <Label>Abre no final de semana?</Label>
-        <Switch
-          thumbColor="#fff"
-          trackColor={{ false: '#ccc', true: '#39CC83' }}
-          value={openOnWeekends}
-          onValueChange={setOpenOnWeekends}
-        />
-      </SwitchContainer>
-
-      <NextButton onPress={handleCreateOrphanage}>
-        <NextButtonText>Cadastrar</NextButtonText>
-      </NextButton>
-    </Container>
-  )
-}
